requirements.txt
starlette[full]
uvicorn
gunicorn
asyncpg
whitenoise
sqlalchemy[asyncio]
pydantic
passlib
psycopg2-binary
python-decouple
PyJWT
Pillow
json2html
aiofiles
aiocsv
websockets
bs4